Better push code

+++ include/Reaching.h
@@ -17,8 +17,8 @@ public:
     MatrixXd ReturnRandomGoalState(MatrixXd X0) override;
     std::vector<MatrixXd> CreateInitOptimisationControls(int horizonLength) override;
 
-    double jointLimsMax[7] = {2.2, 0.6, 2.97, 0, 2.97, 1.5, 2.5};
-    double jointLimsMin[7] = {-2.2, -0.6, -2.97, -1.5, -2.97, -1.1, -2.5};
+    float jointLimsMax[7] = {2.2, 0.6, 2.97, 0, 2.97, 1.5, 2.5};
+    float jointLimsMin[7] = {-2.2, -0.6, -2.97, -1.5, -2.97, -1.1, -2.5};
 
 };
